fix(playback): extract test helpers + kill accessor mutants
Three CI fixes on the Tier-2 follow-up:

1. mod.rs was at 1006 lines after set_scene_active_for_test landed.
   Extracted the two #[cfg(test)] test helpers to a sibling
   test_helpers.rs via #[cfg(test)] mod, mirroring the pattern used
   by recovery.rs and ndi_health.rs.

2. FrameSubmitter::frame_rate_n / frame_rate_d had no direct accessor
   tests, so 6 mutants survived (each → 0, 1, -1). Added two simple
   construction-value tests that kill all 6.

3. PlaybackPipeline::send_command got #[cfg_attr(test, mutants::skip)]
   with a justification comment — it's a one-line wrapper around
   crossbeam_channel::Sender::send; the Ok(()) substitution mutant
   would only be observable by exercising the thread machinery.
